+++ ergo_runtime/src/traits/mod.rs
@@ -11,7 +11,7 @@ use crate as ergo_runtime;
 pub use bind::{bind, bind_error, bind_no_error, Bind, BindImpl};
 pub(crate) use bind::{bind_array, bind_map};
 pub use display::{display, display_any, to_string, Display, DisplayImpl, Formatter};
-pub use into::{into, IntoTyped, IntoTypedImpl};
+pub use into::{into, into_for, into_trait, IntoTyped, IntoTypedImpl};
 pub use nested::{eval_nested, Nested, NestedImpl, NestedValues};
 pub use stored::{
     present_in_store, read_from_store, write_to_store, Stored, StoredContext, StoredImpl,